+++ src/main/java/com/misterfish/interfaces/ServerPlayerInterface.java
@@ -1,8 +1,7 @@
-package com.misterfish.fakes;
+package com.misterfish.interfaces;
 
 import com.misterfish.helper.EntityPlayerActionPack;
 
-public interface ServerPlayerInterface
-{
+public interface ServerPlayerInterface {
     EntityPlayerActionPack getActionPack();
 }