+++ src/main/java/com/misterfish/mixin/PlayerList_offlinePlayersMixin.java
@@ -1,6 +1,5 @@
 package com.misterfish.mixin;
 
-import com.misterfish.OfflinePlayersReworked;
 import com.misterfish.patch.NetHandlerPlayServerFake;
 import com.misterfish.patch.OfflinePlayer;
 import net.minecraft.nbt.CompoundTag;
@@ -16,6 +15,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import static com.misterfish.events.PlayerJoined.playerJoined;
 
 @Mixin(value = PlayerList.class, priority = 900)
 public abstract class PlayerList_offlinePlayersMixin {
@@ -36,7 +37,7 @@ public abstract class PlayerList_offlinePlayersMixin {
             // Replace the network manager with our custom one
             playerIn.connection = new NetHandlerPlayServerFake(this.server, clientConnection, fake, cookie);
         } else {
-            OfflinePlayersReworked.playerJoined(playerIn);
+            playerJoined(playerIn);
         }
     }
 }