+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# OfflinePlayers-Reworked 1.0.9 (MC 1.21.1)
+# OfflinePlayers-Reworked 1.1.0 (MC 1.21.1)
 
 **Update:**
 
-- Fixed dependency issue causing the server to crash on startup
-- Fixed a bug where player would not get killed while `killOnDeath` is `true`.
+- Removed "Other" dependencies
+  
+  OfflinePlayersReworked now doesn't work with a (JsonDB) database or MidnightLib. Instead it uses the Minecraft internal storage as well as SimpleConfig which has no dependencies.